+++ src/app/core/auth/auth.data.ts
@@ -1,6 +1,24 @@
-export const role = ['Director', 'Administrator', 'Editor', 'Staff'];
+export const postions = [
+  ['CS', 'Cardiothoracic Surgeon'],
+  ['AN', 'Anesthesiologist'],
+  ['RN', 'Registered Nurse'],
+  ['CT', 'Cardiothoracic Technician'],
+  ['SN', 'Scrub Nurse'],
+  ['HC', 'Heart Coordinator'],
+  ['RS', 'Researcher'],
+  ['RG', 'Register']
+];
 
-export const menu = {
+export const hospitals = [
+  ['Group1', 'BHT', 'Bangkok Heart Hospital'],
+  ['Group4', 'BCM', 'Bangkok Chiang Mai Hospital']
+];
+
+export const roles = ['Director', 'Administrator', 'Editor', 'Staff'];
+
+export const permissions = ['BDMS', 'Group', 'Hospital'];
+
+export const menus = {
   registry: ['Director', 'Administrator', 'Editor'],
   myPatients: ['Director', 'Administrator', 'Staff'],
   staff: ['Director', 'Administrator'],