+++ src/app/feature/staff/staff.component.ts
@@ -9,6 +9,9 @@ import { Store } from '@ngrx/store';
 import * as fromRoot from '../../app.reducer';
 import * as UI from '../../shared/ui.actions';
 
+import { User } from '../../../app/core/auth/user.model';
+import * as Auth from '../../core/auth/auth.data';
+
 @Component({
   selector: 'app-staff',
   templateUrl: './staff.component.html',
@@ -24,14 +27,26 @@ export class StaffComponent implements OnInit, OnDestroy {
   selectedStaff: Staff;
   staffListSubscription: Subscription;
 
+  user$: Observable<User>;
+  user: User;
+  private userSubscription: Subscription;
+  avHospitals: string[][];
+
   constructor(private router: Router, private staffService: StaffService, private store: Store<fromRoot.State>) {}
 
   ngOnInit() {
+    this.user$ = this.store.select(fromRoot.getUser);
+    this.userSubscription = this.user$.subscribe(user => {
+      this.user = user;
+    });
+
     this.selectedStaff = null;
 
     this.store.dispatch(new UI.StartLoading());
 
-    this.staffListSubscription = this.staffService.getStaffs().subscribe(data => {
+    this.avHospitals = this.getAvailableHospitals();
+
+    this.staffListSubscription = this.staffService.getStaffs(this.avHospitals).subscribe(data => {
       this.dataSource = new MatTableDataSource(data);
       this.dataSource.paginator = this.paginator;
       this.dataSource.sort = this.sort;
@@ -42,6 +57,7 @@ export class StaffComponent implements OnInit, OnDestroy {
 
   ngOnDestroy() {
     this.staffListSubscription.unsubscribe();
+    this.userSubscription.unsubscribe();
   }
 
   applyFilter(filterValue: string) {
@@ -67,5 +83,26 @@ export class StaffComponent implements OnInit, OnDestroy {
       this.staffService.updateStaff(staff);
       this.selectedStaff = null;
     }
+  }
+
+  private getAvailableHospitals(): string[][] {
+    const userHosp = this.user.staff.primaryHospId;
+    const userHospGroup = Auth.hospitals.find(h => h[1] === userHosp)[0];
+    const userPermission = this.user.staff.permission;
+
+    let hospitals: string[][];
+
+    switch (userPermission) {
+      case 'BDMS':
+        hospitals = Auth.hospitals;
+        break;
+      case 'Group':
+        hospitals = Auth.hospitals.filter(h => h[0] === userHospGroup);
+        break;
+      case 'Hospital':
+        hospitals = Auth.hospitals.filter(h => h[1] === userHosp);
+        break;
+    }
+    return hospitals;
   }
 }